Refactor directory model, fix NPEs and resource leaks, add Javadoc
  Introduce LauncherDir and LauncherDirs records to centralize all
  launcher directory paths into the LauncherContext, replacing static
  calls to DirectoryPathUtils.getLauncherDir() throughout the codebase.

  Fixes and improvements:
  - InitProcess: build LauncherDirs in constructor using toURI() path resolution
  - Processes: remove launcherDir field initialization that caused NPE on startup
    before LauncherDirs was set in context
  - Downloads: change context to protected, move launcherDir init into constructor
  - LibrariesDownload: add .path() call on librariesDir(), fix getParentFile()
    null check, add null guard on libraries list
  - FilePathUtils: replace DirectoryPathUtils dependency with LauncherDirs parameter
    in getSelectedVersionJSONPath()
  - VersionJSONParser: add .has() checks before accessing downloads, libraries
    and assetIndex keys, fix misleading error message
  - ManifestParser: wrap Scanner in try-with-resources to prevent resource leak
  - DownloadProgress: add totalSize <= 0 guard, fix percent calculation to avoid
    potential overflow
  - ClientJarInfos: change size field from int to long
  - LauncherUtils: simplify OS fallback to direct return
  - Add Javadoc to LauncherDir and LauncherDirs records

+++ src/main/java/fr/guillaumewlt/model/ClientJarInfos.java
@@ -15,5 +15,5 @@ package fr.guillaumewlt.model;
 public record ClientJarInfos(
         String url,
         String sha1,
-        int size
+        long size
 ) {}
